pybackend/requirements.txt
fastapi 
uvicorn 
langchain
langchain_google_genai
python-dotenv

# uvicorn main:app --host 0.0.0.0 --port $PORT
